[memory] Move all memory stuff into the public API
