+++ include/nova_renderer/memory/allocation_structs.hpp
@@ -1,28 +1,26 @@
 #pragma once
 
-#include "nova_renderer/bytes.hpp"
+#include "nova_renderer/memory/bytes.hpp"
 
-namespace bvestl {
-	namespace polyalloc {
-		/*!
+namespace bvestl::polyalloc {
+    /*!
 		 * Information about an allocation from an allocation strategy
 		 */
-        struct AllocationInfo {
-			/*!
+    struct AllocationInfo {
+        /*!
 			 * \brief The size of the allocation, in bytes
 			 */
-			Bytes size{ 0 };
+        Bytes size{ 0 };
 
-			/*!
+        /*!
 			 * The offset of the allocation, in bytes
 			 */
-			Bytes offset{ 0 };
+        Bytes offset{ 0 };
 
-			/*!
+        /*!
 			 * Data internal to the allocation strategy. For instance, a block allocator might store the block that this allocation corresponds
 			 * to
 			 */
-			void* internal_data = nullptr;
-		};
-	}
+        void* internal_data = nullptr;
+    };
 }
