+++ include/nova_renderer/nova_renderer.hpp
@@ -6,8 +6,8 @@
 #include <string>
 
 #include "nova_renderer/frontend/rendergraph.hpp"
+#include "nova_renderer/memory/polyalloc.hpp"
 #include "nova_renderer/nova_settings.hpp"
-#include "nova_renderer/polyalloc.hpp"
 #include "nova_renderer/renderdoc_app.h"
 #include "nova_renderer/rhi/device_memory_resource.hpp"
 #include "nova_renderer/rhi/forward_decls.hpp"
