+++ include/nova_renderer/rhi/device_memory_resource.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "nova_renderer/allocation_structs.hpp"
-#include "nova_renderer/allocation_strategy.hpp"
-#include "nova_renderer/bytes.hpp"
+#include "nova_renderer/memory//allocation_strategy.hpp"
+#include "nova_renderer/memory//allocation_structs.hpp"
+#include "nova_renderer/memory//bytes.hpp"
 
 using namespace bvestl::polyalloc::operators;
