+++ src/memory/mallocator.cpp
@@ -1,18 +1,14 @@
-#include "mallocator.hpp"
+#include "nova_renderer/memory/mallocator.hpp"
 
-namespace bvestl {
-    namespace polyalloc {
-        void* Mallocator::allocate(const size_t n, int /* flags */) {
-			return std::malloc(n);
-        }
+#include <cstdlib>
 
-        void* Mallocator::allocate(const size_t n, size_t /* alignment */, size_t /* offset */, const int flags) {
-            // TODO: Actually support aligned alloc
-			return allocate(n, flags);
-        }
+namespace bvestl::polyalloc {
+    void* Mallocator::allocate(const size_t n, int /* flags */) { return std::malloc(n); }
 
-        void Mallocator::deallocate(void* p, size_t /* n */) {
-			std::free(p);
-        }
+    void* Mallocator::allocate(const size_t n, size_t /* alignment */, size_t /* offset */, const int flags) {
+        // TODO: Actually support aligned alloc
+        return allocate(n, flags);
     }
-}
+
+    void Mallocator::deallocate(void* p, size_t /* n */) { std::free(p); }
+} // namespace bvestl::polyalloc
