+++ src/render_engine/dx12/dx12_render_engine.hpp
@@ -2,6 +2,8 @@
 
 #include <memory>
 
+// TODO: Don't always use mallocator
+#include "nova_renderer/memory/mallocator.hpp"
 #include "nova_renderer/rhi/render_engine.hpp"
 
 #pragma warning(push, 0)
@@ -13,8 +15,6 @@
 #pragma warning(pop)
 
 #include "dx12_swapchain.hpp"
-// TODO: Not always use mallocator
-#include "../../memory/mallocator.hpp"
 
 // Make sure we don't pollute includees with windows.h problems
 // TODO: Convert to modules and watch this problem _finally_ start to go away
