+++ src/render_objects/procedural_mesh.cpp
@@ -1,11 +1,11 @@
 #include "nova_renderer/frontend/procedural_mesh.hpp"
 
+#include "nova_renderer/memory/block_allocation_strategy.hpp"
+#include "nova_renderer/memory/mallocator.hpp"
 #include "nova_renderer/rhi/device_memory_resource.hpp"
 #include "nova_renderer/rhi/render_engine.hpp"
 #include "nova_renderer/util/logger.hpp"
 
-#include "../memory/block_allocation_strategy.hpp"
-#include "../memory/mallocator.hpp"
 #include "../util/memory_utils.hpp"
 
 using namespace bvestl::polyalloc;
